@@ -64,187 +64,189 @@
     }
   },
   "dependencies": {
-    "@graphql-codegen/core": "^1.11.2",
-    "@graphql-codegen/typescript": "^1.11.2",
-    "apollo-link": "^1.2.2",
-    "apollo-link-ws": "^1.0.8",
-    "brace": "^0.11.1",
-    "deep-equal": "^1.0.1",
+    "@graphql-codegen/core": "1.13.0",
+    "@graphql-codegen/typescript": "1.13.0",
+    "apollo-link": "1.2.13",
+    "apollo-link-ws": "1.0.19",
+    "brace": "0.11.1",
+    "deep-equal": "1.1.1",
     "graphiql": "1.0.0-alpha.0",
     "graphiql-explorer": "0.6.2",
     "graphql": "14.5.8",
-    "graphql-voyager": "^1.0.0-rc.27",
-    "history": "^3.0.0",
-    "hoist-non-react-statics": "^1.0.3",
-    "inflection": "^1.12.0",
-    "invariant": "^2.2.0",
-    "isomorphic-fetch": "^2.2.1",
-    "jsonwebtoken": "^8.5.0",
-    "less": "^3.7.1",
-    "lru-memoize": "^1.0.0",
-    "map-props": "^1.0.0",
-    "match-sorter": "^2.3.0",
-    "multireducer": "^1.0.2",
-    "piping": "^0.3.2",
-    "pretty-error": "^1.2.0",
-    "prop-types": "^15.6.0",
-    "query-string": "^6.1.0",
+    "graphql-voyager": "1.0.0-rc.29",
+    "history": "3.3.0",
+    "hoist-non-react-statics": "1.2.0",
+    "inflection": "1.12.0",
+    "invariant": "2.2.4",
+    "isomorphic-fetch": "2.2.1",
+    "jsonwebtoken": "8.5.1",
+    "less": "3.11.1",
+    "lru-memoize": "1.0.0",
+    "map-props": "1.0.0",
+    "match-sorter": "2.3.0",
+    "multireducer": "1.0.2",
+    "piping": "0.3.2",
+    "pretty-error": "1.2.0",
+    "prop-types": "15.7.2",
+    "query-string": "6.11.1",
     "react": "16.8.6",
-    "react-ace": "^8.0.0",
-    "react-autosuggest": "^9.4.3",
-    "react-bootstrap": "^0.32.1",
-    "react-copy-to-clipboard": "^5.0.0",
+    "react-ace": "8.0.0",
+    "react-autosuggest": "9.4.3",
+    "react-bootstrap": "0.32.4",
+    "react-copy-to-clipboard": "5.0.2",
     "react-dom": "16.8.6",
-    "react-helmet": "^5.2.0",
-    "react-icons": "^3.9.0",
-    "react-modal": "^3.1.2",
-    "react-notification-system": "^0.2.17",
-    "react-notification-system-redux": "^1.2.0",
-    "react-progress-bar-plus": "^1.3.1",
-    "react-redux": "^5.0.6",
-    "react-router": "^3.2.0",
-    "react-router-redux": "^4.0.8",
-    "react-select": "^2.4.3",
-    "react-table": "^6.8.6",
-    "react-tabs": "^2.1.0",
-    "react-toggle": "^4.0.2",
-    "redux": "^4.0.0",
-    "redux-thunk": "^2.2.0",
-    "sanitize-filename": "^1.6.1",
+    "react-helmet": "5.2.1",
+    "react-icons": "3.9.0",
+    "react-modal": "3.11.2",
+    "react-notification-system": "0.2.17",
+    "react-notification-system-redux": "1.2.0",
+    "react-progress-bar-plus": "1.3.1",
+    "react-redux": "5.1.2",
+    "react-router": "3.2.6",
+    "react-router-redux": "4.0.8",
+    "react-select": "2.4.4",
+    "react-table": "6.11.5",
+    "react-tabs": "2.3.1",
+    "react-toggle": "4.1.1",
+    "redux": "4.0.5",
+    "redux-thunk": "2.3.0",
+    "sanitize-filename": "1.6.3",
     "semver": "5.5.1",
-    "styled-components": "^5.0.1",
-    "styled-system": "^5.1.5",
-    "subscriptions-transport-ws": "^0.9.12",
-    "uuid": "^3.0.1",
-    "valid-url": "^1.0.9"
+    "styled-components": "5.0.1",
+    "styled-system": "5.1.5",
+    "subscriptions-transport-ws": "0.9.16",
+    "uuid": "3.4.0",
+    "valid-url": "1.0.9"
   },
   "devDependencies": {
-    "@babel/core": "^7.3.3",
-    "@babel/plugin-proposal-class-properties": "^7.3.3",
-    "@babel/plugin-proposal-decorators": "^7.3.0",
-    "@babel/plugin-proposal-do-expressions": "^7.0.0",
-    "@babel/plugin-proposal-export-default-from": "^7.0.0",
-    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
-    "@babel/plugin-proposal-function-bind": "^7.0.0",
-    "@babel/plugin-proposal-function-sent": "^7.2.0",
-    "@babel/plugin-proposal-json-strings": "^7.2.0",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
-    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
-    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-syntax-import-meta": "^7.2.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.8.3",
-    "@babel/register": "^7.0.0",
-    "@babel/runtime": "^7.0.0",
-    "@types/clean-webpack-plugin": "^0.1.3",
-    "@types/concurrently": "^5.1.0",
-    "@types/cypress": "^1.1.3",
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.3",
-    "@types/express-session": "^1.17.0",
-    "@types/extract-text-webpack-plugin": "^3.0.4",
-    "@types/file-loader": "^4.2.0",
-    "@types/fork-ts-checker-webpack-plugin": "^0.4.5",
-    "@types/jquery": "^3.3.33",
-    "@types/lodash": "^4.14.149",
-    "@types/mini-css-extract-plugin": "^0.9.1",
-    "@types/node-sass": "^4.11.0",
-    "@types/optimize-css-assets-webpack-plugin": "^5.0.1",
-    "@types/react": "^16.9.23",
-    "@types/react-addons-test-utils": "^0.14.25",
-    "@types/react-dom": "^16.9.5",
-    "@types/react-helmet": "^5.0.15",
-    "@types/react-hot-loader": "^4.1.1",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router": "^5.1.4",
-    "@types/react-router-redux": "^5.0.18",
-    "@types/react-toggle": "^4.0.2",
-    "@types/redux-devtools": "^3.0.47",
-    "@types/redux-devtools-dock-monitor": "^1.1.33",
-    "@types/redux-devtools-log-monitor": "^1.0.34",
-    "@types/redux-logger": "^3.0.7",
-    "@types/sinon": "^7.5.2",
-    "@types/styled-components": "^5.1.0",
-    "@types/styled-system": "^5.1.9",
-    "@types/terser-webpack-plugin": "^2.2.0",
-    "@types/unused-files-webpack-plugin": "^3.4.1",
-    "@types/webpack": "^4.41.7",
-    "@types/webpack-bundle-analyzer": "^2.13.3",
-    "@types/webpack-dev-middleware": "^3.7.0",
-    "@types/webpack-hot-middleware": "^2.25.0",
-    "@types/ws": "^7.2.3",
-    "@typescript-eslint/eslint-plugin": "^2.24.0",
-    "@typescript-eslint/parser": "^2.24.0",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.0",
-    "babel-plugin-istanbul": "^5.1.1",
-    "babel-plugin-styled-components": "^1.10.6",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.10",
-    "babel-plugin-typecheck": "^2.0.0",
-    "better-npm-run": "^0.1.0",
-    "bootstrap-loader": "^2.2.0",
-    "bootstrap-sass": "^3.3.7",
-    "clean-webpack-plugin": "^0.1.17",
-    "concurrently": "^3.5.0",
-    "css-loader": "^0.28.11",
-    "cypress": "^3.2.0",
-    "dotenv": "^5.0.1",
-    "eslint": "^6.5.1",
+    "@babel/core": "7.9.0",
+    "@babel/plugin-proposal-class-properties": "7.8.3",
+    "@babel/plugin-proposal-decorators": "7.8.3",
+    "@babel/plugin-proposal-do-expressions": "7.8.3",
+    "@babel/plugin-proposal-export-default-from": "7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "7.8.3",
+    "@babel/plugin-proposal-function-bind": "7.8.3",
+    "@babel/plugin-proposal-function-sent": "7.8.3",
+    "@babel/plugin-proposal-json-strings": "7.8.3",
+    "@babel/plugin-proposal-logical-assignment-operators": "7.8.3",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "7.8.3",
+    "@babel/plugin-proposal-numeric-separator": "7.8.3",
+    "@babel/plugin-proposal-optional-chaining": "7.8.3",
+    "@babel/plugin-proposal-pipeline-operator": "7.8.3",
+    "@babel/plugin-proposal-throw-expressions": "7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "7.8.3",
+    "@babel/plugin-syntax-import-meta": "7.8.3",
+    "@babel/preset-env": "7.9.5",
+    "@babel/preset-react": "7.8.3",
+    "@babel/preset-typescript": "7.8.3",
+    "@babel/register": "7.8.6",
+    "@babel/runtime": "7.8.7",
+    "@types/clean-webpack-plugin": "0.1.3",
+    "@types/concurrently": "5.1.0",
+    "@types/cypress": "1.1.3",
+    "@types/dotenv": "8.2.0",
+    "@types/express": "4.17.3",
+    "@types/express-session": "1.17.0",
+    "@types/extract-text-webpack-plugin": "3.0.4",
+    "@types/file-loader": "4.2.0",
+    "@types/fork-ts-checker-webpack-plugin": "0.4.5",
+    "@types/isomorphic-fetch": "0.0.35",
+    "@types/jquery": "3.3.33",
+    "@types/lodash": "4.14.149",
+    "@types/mini-css-extract-plugin": "0.9.1",
+    "@types/node-sass": "4.11.0",
+    "@types/optimize-css-assets-webpack-plugin": "5.0.1",
+    "@types/react": "16.9.23",
+    "@types/react-addons-test-utils": "0.14.25",
+    "@types/react-bootstrap": "0.32.21",
+    "@types/react-dom": "16.9.5",
+    "@types/react-helmet": "5.0.15",
+    "@types/react-hot-loader": "4.1.1",
+    "@types/react-redux": "7.1.7",
+    "@types/react-router": "^3.0.8",
+    "@types/react-router-redux": "4.0.44",
+    "@types/react-toggle": "4.0.2",
+    "@types/redux-devtools": "3.0.47",
+    "@types/redux-devtools-dock-monitor": "1.1.33",
+    "@types/redux-devtools-log-monitor": "1.0.34",
+    "@types/redux-logger": "3.0.7",
+    "@types/sinon": "7.5.2",
+    "@types/styled-components": "5.1.0",
+    "@types/styled-system": "5.1.9",
+    "@types/terser-webpack-plugin": "2.2.0",
+    "@types/unused-files-webpack-plugin": "3.4.1",
+    "@types/webpack": "4.41.7",
+    "@types/webpack-bundle-analyzer": "2.13.3",
+    "@types/webpack-dev-middleware": "3.7.0",
+    "@types/webpack-hot-middleware": "2.25.0",
+    "@types/ws": "7.2.3",
+    "@typescript-eslint/eslint-plugin": "2.24.0",
+    "@typescript-eslint/parser": "2.24.0",
+    "babel-eslint": "10.1.0",
+    "babel-loader": "8.0.6",
+    "babel-plugin-istanbul": "5.2.0",
+    "babel-plugin-styled-components": "1.10.7",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.24",
+    "babel-plugin-typecheck": "2.0.1",
+    "better-npm-run": "0.1.1",
+    "bootstrap-loader": "2.2.0",
+    "bootstrap-sass": "3.4.1",
+    "clean-webpack-plugin": "0.1.19",
+    "concurrently": "3.6.1",
+    "css-loader": "0.28.11",
+    "cypress": "3.8.3",
+    "dotenv": "5.0.1",
+    "eslint": "6.8.0",
     "eslint-config-airbnb": "16.1.0",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-chai-friendly": "^0.4.1",
-    "eslint-plugin-cypress": "^2.0.1",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.9.1",
-    "express": "^4.13.3",
-    "express-session": "^1.12.1",
+    "eslint-config-prettier": "6.10.0",
+    "eslint-loader": "1.9.0",
+    "eslint-plugin-chai-friendly": "0.4.1",
+    "eslint-plugin-cypress": "2.10.3",
+    "eslint-plugin-import": "2.18.2",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.19.0",
+    "express": "4.17.1",
+    "express-session": "1.17.0",
     "extract-hoc": "0.0.5",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.11",
-    "font-awesome": "^4.7.0",
+    "extract-text-webpack-plugin": "3.0.2",
+    "file-loader": "1.1.11",
+    "font-awesome": "4.7.0",
     "font-awesome-webpack": "0.0.4",
-    "fork-ts-checker-webpack-plugin": "^4.0.2",
-    "husky": "^4.2.3",
-    "ignore-loader": "^0.1.2",
-    "jquery": "^3.4.1",
-    "json-loader": "^0.5.4",
-    "less-loader": "^4.1.0",
-    "lint-staged": "^6.1.1",
-    "mini-css-extract-plugin": "^0.4.0",
-    "node-sass": "^4.13.0",
-    "nyc": "^13.3.0",
-    "optimize-css-assets-webpack-plugin": "^4.0.2",
+    "fork-ts-checker-webpack-plugin": "4.0.5",
+    "husky": "4.2.3",
+    "ignore-loader": "0.1.2",
+    "jquery": "3.4.1",
+    "json-loader": "0.5.7",
+    "less-loader": "4.1.0",
+    "lint-staged": "6.1.1",
+    "mini-css-extract-plugin": "0.4.5",
+    "node-sass": "4.13.1",
+    "nyc": "13.3.0",
+    "optimize-css-assets-webpack-plugin": "4.0.3",
     "prettier": "1.19.1",
-    "react-a11y": "^0.2.6",
-    "react-addons-test-utils": "^15.0.3",
-    "react-hot-loader": "^4.6.5",
-    "redux-devtools": "^3.4.1",
-    "redux-devtools-dock-monitor": "^1.1.2",
-    "redux-devtools-log-monitor": "^1.3.0",
+    "react-a11y": "0.2.8",
+    "react-addons-test-utils": "15.6.2",
+    "react-hot-loader": "4.12.9",
+    "redux-devtools": "3.5.0",
+    "redux-devtools-dock-monitor": "1.1.3",
+    "redux-devtools-log-monitor": "1.4.0",
     "redux-logger": "3.0.1",
-    "resolve-url-loader": "^2.3.0",
-    "sass-loader": "^7.0.1",
-    "sinon": "^1.17.7",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^0.20.3",
-    "terser-webpack-plugin": "^2.3.3",
+    "resolve-url-loader": "2.3.2",
+    "sass-loader": "7.3.1",
+    "sinon": "1.17.7",
+    "source-map-loader": "0.2.4",
+    "style-loader": "0.20.3",
+    "terser-webpack-plugin": "2.3.5",
     "timekeeper": "1.0.0",
-    "typescript": "^3.8.3",
-    "unused-files-webpack-plugin": "^3.4.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.14.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-middleware": "^3.1.3",
-    "webpack-hot-middleware": "^2.22.2",
-    "webpack-isomorphic-tools": "^3.0.5"
+    "typescript": "3.8.3",
+    "unused-files-webpack-plugin": "3.4.0",
+    "url-loader": "1.1.2",
+    "webpack": "4.42.0",
+    "webpack-bundle-analyzer": "3.6.0",
+    "webpack-cli": "3.3.11",
+    "webpack-dev-middleware": "3.7.2",
+    "webpack-hot-middleware": "2.25.0",
+    "webpack-isomorphic-tools": "3.0.6"
   },
   "engines": {
     "node": ">=8.9.1"
