@@ -1,9 +1,7 @@
-import "regenerator-runtime/runtime";
-const {
-  sdl,
-  actionsCodegen
-} = require('./services');
+import 'regenerator-runtime/runtime';
+
 const fs = require('fs');
+const { sdl, actionsCodegen } = require('./services');
 const { getFlagValue, OUTPUT_FILE_FLAG } = require('./utils/commandUtils');
 
 const commandArgs = process.argv;
@@ -12,10 +10,12 @@ const outputFilePath = getFlagValue(commandArgs, OUTPUT_FILE_FLAG);
 const logOutput = (log) => {
   try {
     fs.writeFile(outputFilePath, log, 'utf8', () => {
-      console.log(JSON.stringify({
-        success: true,
-        output_file_path: outputFilePath
-      }));
+      console.log(
+        JSON.stringify({
+          success: true,
+          output_file_path: outputFilePath,
+        }),
+      );
     });
   } catch (e) {
     console.error(`could not write output to "${outputFilePath}"`);
@@ -25,7 +25,7 @@ const logOutput = (log) => {
 
 const handleArgs = () => {
   const rootArg = commandArgs[2];
-  switch(rootArg) {
+  switch (rootArg) {
     case 'sdl':
       const sdlSubCommands = commandArgs.slice(3);
       return sdl(sdlSubCommands);
@@ -33,25 +33,26 @@ const handleArgs = () => {
       const actionCodegenSubCommands = commandArgs.slice(3);
       return actionsCodegen(actionCodegenSubCommands);
     default:
-      return;
   }
-}
+};
 
 try {
-  let cliResponse = handleArgs();
+  const cliResponse = handleArgs();
   if (cliResponse.error) {
-    throw Error(cliResponse.error)
+    throw Error(cliResponse.error);
   }
   if (cliResponse.constructor.name === 'Promise') {
-    cliResponse.then(r => {
-      if (r.error) {
-        throw Error(r.error)
-      }
-      logOutput(JSON.stringify(r));
-    }).catch(e => {
-      console.error(e);
-      process.exit(1);
-    })
+    cliResponse
+      .then((r) => {
+        if (r.error) {
+          throw Error(r.error);
+        }
+        logOutput(JSON.stringify(r));
+      })
+      .catch((e) => {
+        console.error(e);
+        process.exit(1);
+      });
   } else {
     logOutput(JSON.stringify(cliResponse));
   }
