@@ -1,27 +1,35 @@
 const fetch = require('node-fetch');
-const { getTemplatePath } = require('../../utils/utils');
-const { parseCustomTypes, getActionTypes } = require('../../shared/utils/hasuraCustomTypeUtils')
-const { getFrameworkCodegen } = require('./template');
-const { getActionDefinitionSdl, getTypesSdl } = require('../../shared/utils/sdlUtils');
 const { parse: sdlParse } = require('graphql/language/parser');
+const { getTemplatePath } = require('../../utils/utils');
+const {
+  parseCustomTypes,
+  getActionTypes,
+} = require('../../shared/utils/hasuraCustomTypeUtils');
+const { getFrameworkCodegen } = require('./template');
+const {
+  getActionDefinitionSdl,
+  getTypesSdl,
+} = require('../../shared/utils/sdlUtils');
 
 const getActionsCodegen = async (payload) => {
-
   const {
     action_name: actionName,
-    sdl: {
-      complete: sdlComplete
-    },
+    sdl: { complete: sdlComplete },
     derive,
-    codegen_config: codegenConfig
+    codegen_config: codegenConfig,
   } = payload;
 
   try {
-    const codegenResp = await getFrameworkCodegen(actionName, sdlComplete, derive, codegenConfig)
+    const codegenResp = await getFrameworkCodegen(
+      actionName,
+      sdlComplete,
+      derive,
+      codegenConfig,
+    );
     if (codegenResp.error) {
-      throw Error(codegenResp.error)
+      throw Error(codegenResp.error);
     } else {
-      return codegenResp.files
+      return codegenResp.files;
     }
   } catch (e) {
     throw e;
@@ -29,6 +37,5 @@ const getActionsCodegen = async (payload) => {
 };
 
 module.exports = {
-  getActionsCodegen 
+  getActionsCodegen,
 };
-
