@@ -4,19 +4,19 @@ const typescriptPlugin = require('@graphql-codegen/typescript');
 const { camelize } = require('inflection');
 
 const fetch = require('node-fetch');
-const path = require('path')
+const path = require('path');
 const fs = require('fs');
-const { getTemplatePath } = require('../../utils/utils')
+const { getTemplatePath } = require('../../utils/utils');
 
 const CODEGENERATOR_NOT_FOUND = 'given codegen framework not found';
 const FILE_SYSTEM_PATH = 'fs_path';
 const URL_PATH = 'url path';
 
 const resolveCodegeneratorPath = (codegenConfig) => {
-  let { framework } = codegenConfig;
+  const { framework } = codegenConfig;
   let codegeneratorPath = codegenConfig.uri;
   if (!codegeneratorPath) {
-    codegeneratorPath = getTemplatePath(framework)
+    codegeneratorPath = getTemplatePath(framework);
   }
   return codegeneratorPath;
 };
@@ -28,7 +28,7 @@ const resolveCodegeneratorFromUrl = async (url) => {
     if (fetchResp.status >= 300) {
       throw Error(CODEGENERATOR_NOT_FOUND);
     }
-    const codegeneratorText = await fetchResp.text()
+    const codegeneratorText = await fetchResp.text();
     eval(`${codegeneratorText} codegenerator = templater`);
     return codegenerator;
   } catch (e) {
@@ -39,7 +39,9 @@ const resolveCodegeneratorFromUrl = async (url) => {
 const resolveCodegeneratorFromFs = async (fsPath) => {
   let codegenerator;
   try {
-    const codegeneratorText = fs.readFileSync(path.resolve(fsPath), { encoding: 'utf8'});
+    const codegeneratorText = fs.readFileSync(path.resolve(fsPath), {
+      encoding: 'utf8',
+    });
     eval(`${codegeneratorText}\n codegenerator = templater`);
     return codegenerator;
   } catch (e) {
@@ -48,23 +50,22 @@ const resolveCodegeneratorFromFs = async (fsPath) => {
 };
 
 const resolveCodegenerator = async (codegenConfig) => {
-
   const codegeneratorPath = resolveCodegeneratorPath(codegenConfig);
   if (!codegeneratorPath) {
-    throw Error(CODEGENERATOR_NOT_FOUND)
+    throw Error(CODEGENERATOR_NOT_FOUND);
   }
 
-  let codegenerator
+  let codegenerator;
   let pathType = URL_PATH;
   try {
-    new URL(codegeneratorPath)
+    new URL(codegeneratorPath);
   } catch (_) {
     pathType = FILE_SYSTEM_PATH;
   }
 
   try {
     if (pathType === FILE_SYSTEM_PATH) {
-      codegenerator = await resolveCodegeneratorFromFs(codegeneratorPath)
+      codegenerator = await resolveCodegeneratorFromFs(codegeneratorPath);
     } else {
       codegenerator = await resolveCodegeneratorFromUrl(codegeneratorPath);
     }
@@ -73,14 +74,17 @@ const resolveCodegenerator = async (codegenConfig) => {
   }
 
   return codegenerator;
+};
 
-}
-
-const getCodegenFiles = async (actionName, actionsSdl, derive, codegenConfig) => {
-
+const getCodegenFiles = async (
+  actionName,
+  actionsSdl,
+  derive,
+  codegenConfig,
+) => {
   let codegenerator;
   try {
-    codegenerator = await resolveCodegenerator(codegenConfig)
+    codegenerator = await resolveCodegenerator(codegenConfig);
   } catch (e) {
     throw e;
   }
@@ -91,24 +95,31 @@ const getCodegenFiles = async (actionName, actionsSdl, derive, codegenConfig) =>
   }
 
   return codegenFiles;
+};
 
-}
-
-const getFrameworkCodegen = async (actionName, actionsSdl, derive, codegenConfig) => {
-
+const getFrameworkCodegen = async (
+  actionName,
+  actionsSdl,
+  derive,
+  codegenConfig,
+) => {
   try {
-    const codegenFiles = await getCodegenFiles(actionName, actionsSdl, derive, codegenConfig);
+    const codegenFiles = await getCodegenFiles(
+      actionName,
+      actionsSdl,
+      derive,
+      codegenConfig,
+    );
     return {
-      files: codegenFiles
-    }
+      files: codegenFiles,
+    };
   } catch (e) {
     return {
-      error: e.message
-    }
+      error: e.message,
+    };
   }
-
 };
 
 module.exports = {
-  getFrameworkCodegen
+  getFrameworkCodegen,
 };
