@@ -3,5 +3,5 @@ const sdlHandler = require('./sdl');
 
 module.exports = {
   actionsCodegen: actionsCodegenHandler,
-  sdl: sdlHandler
+  sdl: sdlHandler,
 };
