@@ -1,5 +1,5 @@
 const handler = require('./handler');
-const { getInputPayload } = require('../../../utils/commandUtils')
+const { getInputPayload } = require('../../../utils/commandUtils');
 
 const command = (subCommands) => {
   const rootInput = subCommands[0];
