@@ -1,23 +1,28 @@
-const { getAllActionsFromSdl, getAllTypesFromSdl } = require('../../../shared/utils/sdlUtils');
-const { reformCustomTypes } = require('../../../shared/utils/hasuraCustomTypeUtils')
+const {
+  getAllActionsFromSdl,
+  getAllTypesFromSdl,
+} = require('../../../shared/utils/sdlUtils');
+const {
+  reformCustomTypes,
+} = require('../../../shared/utils/hasuraCustomTypeUtils');
 
 const handlePayload = (payload) => {
-
   const response = {
     body: null,
-    status: 200
+    status: 200,
   };
 
   const { sdl } = payload;
 
-  let customTypes, typesParseError;
+  let customTypes;
+  let typesParseError;
 
   if (!sdl.complete.trim()) {
     response.body = {
       actions: [],
-      types: reformCustomTypes([])
-    }
-    return response
+      types: reformCustomTypes([]),
+    };
+    return response;
   }
 
   try {
@@ -25,17 +30,17 @@ const handlePayload = (payload) => {
   } catch (e) {
     typesParseError = e;
   }
-  
 
   if (typesParseError) {
     response.body = {
-      error: typesParseError.message
+      error: typesParseError.message,
     };
     response.status = 400;
     return response;
   }
 
-  let allActions, actionsParseError;
+  let allActions;
+  let actionsParseError;
   try {
     allActions = getAllActionsFromSdl(sdl.complete);
   } catch (e) {
@@ -44,38 +49,24 @@ const handlePayload = (payload) => {
 
   if (actionsParseError) {
     response.body = {
-      error: actionsParseError.message
+      error: actionsParseError.message,
     };
     response.status = 400;
   }
 
   response.body = {
     actions: allActions,
-    types: customTypes
+    types: customTypes,
   };
 
   return response;
-
-}
+};
 
 const requestHandler = (payload) => {
-
-  const {
-    body, status
-  } = handlePayload(payload)
+  const { body, status } = handlePayload(payload);
 
   return body;
-
 };
 
 module.exports = requestHandler;
 module.exports.handlePayload = handlePayload;
-
-
-
-
-
-
-
-
-
