@@ -1,29 +1,38 @@
-const { getTypesSdl, getActionDefinitionSdl } = require('../../../shared/utils/sdlUtils');
+const {
+  getTypesSdl,
+  getActionDefinitionSdl,
+} = require('../../../shared/utils/sdlUtils');
 const deriveAction = require('../../../shared/utils/deriveAction').default;
 
 const handlePayload = (payload) => {
-
   const { actions, types, derive } = payload;
   const {
     operation: toDeriveOperation,
     introspection_schema: introspectionSchema,
-    action_name: actionName
+    action_name: actionName,
   } = derive || {};
 
   const response = {
     body: null,
-    status: 200
+    status: 200,
   };
 
   let actionSdl = '';
   let typesSdl = '';
-  let actionSdlError, typesSdlError, deriveActionError;
+  let actionSdlError;
+  let typesSdlError;
+  let deriveActionError;
 
   if (actions) {
     try {
-      actions.forEach(a => {
-        actionSdl += getActionDefinitionSdl(a.name, a.definition.type, a.definition.arguments, a.definition.output_type) + '\n';
-      })
+      actions.forEach((a) => {
+        actionSdl += `${getActionDefinitionSdl(
+          a.name,
+          a.definition.type,
+          a.definition.arguments,
+          a.definition.output_type,
+        )}\n`;
+      });
     } catch (e) {
       actionSdlError = e;
     }
@@ -41,10 +50,19 @@ const handlePayload = (payload) => {
 
   if (toDeriveOperation) {
     try {
-      const derivation = deriveAction(toDeriveOperation, introspectionSchema, actionName);
-      const derivedActionSdl = getActionDefinitionSdl(derivation.action.name, derivation.action.type, derivation.action.arguments, derivation.action.output_type);
+      const derivation = deriveAction(
+        toDeriveOperation,
+        introspectionSchema,
+        actionName,
+      );
+      const derivedActionSdl = getActionDefinitionSdl(
+        derivation.action.name,
+        derivation.action.type,
+        derivation.action.arguments,
+        derivation.action.output_type,
+      );
       const derivedTypesSdl = getTypesSdl(derivation.types);
-      sdl = `${derivedActionSdl}\n\n${derivedTypesSdl}\n\n${sdl}`
+      sdl = `${derivedActionSdl}\n\n${derivedTypesSdl}\n\n${sdl}`;
     } catch (e) {
       deriveActionError = e;
     }
@@ -52,7 +70,7 @@ const handlePayload = (payload) => {
 
   if (actionSdlError) {
     response.body = {
-      error: 'invalid actions definition'
+      error: 'invalid actions definition',
     };
     response.status = 400;
     return response;
@@ -60,7 +78,7 @@ const handlePayload = (payload) => {
 
   if (deriveActionError) {
     response.body = {
-      error: `could not derive action: ${deriveActionError.message}`
+      error: `could not derive action: ${deriveActionError.message}`,
     };
     response.status = 400;
     return response;
@@ -68,7 +86,7 @@ const handlePayload = (payload) => {
 
   if (typesSdlError) {
     response.body = {
-      error: 'invalid types'
+      error: 'invalid types',
     };
     response.status = 400;
     return response;
@@ -76,23 +94,18 @@ const handlePayload = (payload) => {
 
   response.body = {
     sdl: {
-      complete: sdl
-    }
+      complete: sdl,
+    },
   };
 
   return response;
-
-}
+};
 
 const requestHandler = (payload) => {
-
-  const {
-    body, status
-  } = handlePayload(payload)
+  const { body, status } = handlePayload(payload);
 
   return body;
-
-}
+};
 
 module.exports = requestHandler;
 module.exports.handlePayload = handlePayload;
