@@ -4,14 +4,12 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"sort"
 	"strings"
 
 	"github.com/hasura/graphql-engine/cli"
-	"github.com/markbates/pkger"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/cobra/doc"
@@ -105,26 +103,12 @@ func genReSTCustom(cmd *cobra.Command, w io.Writer, titlePrefix string, linkHand
 	buf := new(bytes.Buffer)
 	name := cmd.CommandPath()
 	ref := strings.Replace(name, " ", "_", -1)
-	cliDocPath := "manifests/docs/" + ref + ".rst"
 	short := cmd.Short
 	long := cmd.Long
 	if len(long) == 0 {
 		long = short
 	}
-	file, err := pkger.Open(cliDocPath)
-	if err != nil {
-		return err
-	}
-	fileInfo, err := ioutil.ReadAll(file)
-	if err != nil {
-		return err
-	}
-	var info string
-	if err != nil || string(fileInfo) == "" {
-		info = short
-	} else {
-		info = string(fileInfo)
-	}
+	info := short
 
 	buf.WriteString(".. _" + ref + ":\n\n")
 
@@ -184,7 +168,7 @@ func genReSTCustom(cmd *cobra.Command, w io.Writer, titlePrefix string, linkHand
 	if !cmd.DisableAutoGenTag {
 		buf.WriteString("*Auto generated by spf13/cobra*\n")
 	}
-	_, err = buf.WriteTo(w)
+	_, err := buf.WriteTo(w)
 	return err
 }
 
