@@ -8,6 +8,7 @@ package migrate
 import (
 	"bytes"
 	"container/list"
+	"crypto/tls"
 	"fmt"
 	"io"
 	"os"
@@ -102,7 +103,7 @@ type Migrate struct {
 
 // New returns a new Migrate instance from a source URL and a database URL.
 // The URL scheme is defined by each driver.
-func New(sourceUrl string, databaseUrl string, cmd bool, configVersion int, logger *log.Logger) (*Migrate, error) {
+func New(sourceUrl string, databaseUrl string, cmd bool, configVersion int, tlsConfig *tls.Config, logger *log.Logger) (*Migrate, error) {
 	m := newCommon(cmd)
 
 	sourceName, err := schemeFromUrl(sourceUrl)
@@ -136,7 +137,7 @@ func New(sourceUrl string, databaseUrl string, cmd bool, configVersion int, logg
 		m.sourceDrv.DefaultParser(source.DefaultParsev2)
 	}
 
-	databaseDrv, err := database.Open(databaseUrl, cmd, logger)
+	databaseDrv, err := database.Open(databaseUrl, cmd, tlsConfig, logger)
 	if err != nil {
 		log.Debug(err)
 		return nil, err
