@@ -14,7 +14,7 @@ const winMoveMatrix = [
   [3, 4, 5],
   [6, 7, 8],
   [0, 3, 6],
-  [1, 4, 8],
+  [1, 4, 7],
   [2, 5, 8],
   [0, 4, 8],
   [2, 4, 6],
