@@ -6,7 +6,10 @@ const babelrc = fs.readFileSync('.babelrc');
 let config;
 
 try {
-  config = JSON.parse(babelrc);
+  config = {
+    ...JSON.parse(babelrc),
+    extensions: ['.ts', '.tsx', '.js', '.jsx'],
+  };
 } catch (err) {
   console.error('==>     ERROR: Error parsing your .babelrc.');
   console.error(err);
