@@ -1,5 +1,5 @@
+import 'react-toggle/style.css';
 import Toggle from 'react-toggle';
 import './Toggle.css';
-import 'react-toggle/style.css';
 
 export default Toggle;
