@@ -3,7 +3,7 @@ import { parse as sdlParse } from 'graphql/language/parser';
 import styles from './Styles.scss';
 import Tooltip from './Tooltip';
 import CrossIcon from '../../../../Common/Icons/Cross';
-import SDLEditor from '../../../../Common/AceEditor/SDLEditor';
+import AceEditor from '../../../../Common/AceEditor/BaseEditor';
 
 const editorLabel = 'Action definition';
 const editorTooltip =
@@ -73,13 +73,14 @@ const ActionDefinitionEditor = ({
             </div>
           )}
         </div>
-        <SDLEditor
+        <AceEditor
           name="sdl-editor"
           value={value}
           onChange={onChangeWithError}
           placeholder={placeholder}
           markers={markers}
           height="200px"
+          mode="graphqlschema"
           width="600px"
         />
       </div>
