@@ -3,4 +3,5 @@ import globals from '../../../Globals';
 export const appPrefix = '/actions';
 export const pageTitle = 'Actions';
 export const CODEGEN_REPO = 'hasura/codegen-assets';
-export const ALL_FRAMEWORKS_FILE_PATH = `${globals.assetsPath}/common/codegen/frameworks.json`;
+export const BASE_CODEGEN_PATH = `${globals.assetsPath}/common/codegen`;
+export const ALL_FRAMEWORKS_FILE_PATH = `${BASE_CODEGEN_PATH}/frameworks.json`;
