@@ -26,7 +26,7 @@ export const TypedInput = ({
   const isAutoIncrement = isColumnAutoIncrement(col);
   const placeHolder = hasDefault ? colDefault : getPlaceholder(colType);
   const getDefaultValue = () => {
-    if (prevValue) return prevValue;
+    if (prevValue !== undefined) return prevValue;
     if (clone && colName in clone) return clone[colName];
     return '';
   };
@@ -109,7 +109,7 @@ export const TypedInput = ({
 
     case BOOLEAN:
       return (
-        <select {...standardInputProps} defaultValue={placeHolder}>
+        <select {...standardInputProps}>
           <option value="" disabled>
             -- bool --
           </option>
