@@ -18,7 +18,7 @@ class TestGraphQLQueryBasic:
 
     def test_select_query_author_with_include_directive(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + '/select_query_author_include_directive.yaml', transport)
-    
+
     # Can't run server upgrade tests, as this test has a schema change
     @pytest.mark.skip_server_upgrade_test
     def test_select_various_postgres_types(self, hge_ctx, transport):
@@ -388,7 +388,7 @@ class TestGraphQLQueryBoolExpJsonB:
     def dir(cls):
         return 'queries/graphql_query/boolexp/jsonb'
 
-@pytest.mark.parametrize("transport", ['http', 'websocket'])
+@pytest.mark.parametrize("transport", ['http', 'websocket', 'subscription'])
 @usefixtures('per_class_tests_db_state')
 class TestGraphQLQueryBoolExpPostGIS:
 
@@ -606,3 +606,6 @@ class TestGraphQLQueryCaching:
 
     def test_include_directive(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + '/include_directive.yaml', transport)
+
+    def test_introspection(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/introspection.yaml', transport)
